backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0  # SQLite async driver
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.13.1  # Database migrations

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.26.0

# Configuration
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2

# Async task scheduling (for future batch processing)
apscheduler==3.10.4
